# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  util.Res:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:80
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /plugin/create:
    post:
      consumes:
      - application/json
      description: Create Plugin
      operationId: CreatePlugin
      parameters:
      - description: Create Plugin
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/index.Plugin'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: CreatePlugin
      tags:
      - CreatePlugin
  /plugin/delete:
    delete:
      consumes:
      - application/json
      description: Delete Plugin
      operationId: DeletePlugin
      parameters:
      - description: Plugin_Name
        in: query
        name: plugin_name
        required: true
        type: string
      - description: Plugin Version
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: DeletePlugin
      tags:
      - DeletePlugin
  /plugin/list:
    get:
      consumes:
      - application/json
      description: List Plugin
      operationId: ListPlugin
      parameters:
      - description: Plugin Name
        in: query
        name: plugin_name
        type: string
      - description: Plugin Version
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: ListPlugin
      tags:
      - ListPlugin
  /plugin/listall:
    get:
      consumes:
      - application/json
      description: List Plugin ALL Version
      operationId: ListPluginAllVersion
      parameters:
      - description: Plugin Name
        in: query
        name: plugin_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: ListPluginAllVersion
      tags:
      - ListPluginAllVersion
  /plugin/search:
    get:
      consumes:
      - application/json
      description: Search Plugin
      operationId: SearchPlugin
      parameters:
      - description: Plugin Name
        in: query
        name: plugin_name
        type: string
      - description: Plugin Version
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: SearchPlugin
      tags:
      - SearchPlugin
  /plugin/searchall:
    get:
      consumes:
      - application/json
      description: Search Plugin ALL Version
      operationId: SearchPluginAllVersion
      parameters:
      - description: Plugin Name
        in: query
        name: plugin_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: SearchPluginAllVersion
      tags:
      - SearchPluginAllVersion
  /plugin/update:
    put:
      consumes:
      - application/json
      description: Update Plugin
      operationId: UpdatePlugin
      parameters:
      - description: Update Plugin
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/index.Plugin'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: UpdatePlugin
      tags:
      - UpdatePlugin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"